Add RigidBody2D to Detector
Makes collision detection between Detectors and Detectees much more
consistent as it doesn't have to worry about adding one at runtime.
The Detector is meant to be a nested object so it's RigidBody2D
does not interfere with one in the attached object. This was tested,
however it is currently unknown which RigidBody is used for collision
detection in the case that the object and the nested object both have
one, though it may not matter.